code/backend/requirements.txt
# onlyGen Backend — Python Dependencies
# Install with: pip install -r requirements.txt

# ── Core Framework ────────────────────────────────────────────
fastapi>=0.110.0
uvicorn[standard]>=0.27.0

# ── Google DeepMind — LLM + Agent Development Kit ────────────
google-adk>=1.25.1
google-genai>=1.55.0

# ── Database ──────────────────────────────────────────────────
aiosqlite>=0.20.0

# ── Data Validation ───────────────────────────────────────────
pydantic>=2.0.0

# ── Environment ───────────────────────────────────────────────
python-dotenv>=1.0.0

# ── Observability ─────────────────────────────────────────────

# ── HTTP Client (for integrations) ───────────────────────────
httpx>=0.27.0

# ── Airia — Enterprise Orchestration & AI Gateway ────────────

# ── ElevenLabs — Voice Agent & Text-to-Speech ────────────────

# ── Structured logging (used by Datadog metrics helper) ───────
structlog>=24.1.0

# ── Testing ───────────────────────────────────────────────────
pytest>=8.0.0
pytest-asyncio>=0.23.0
